Split out ffma_thread_cache out of ffma
