+++ src/memory_allocator/ffma.c
@@ -40,6 +40,7 @@
 #include "hugepage_cache.h"
 
 #include "ffma.h"
+#include "ffma_thread_cache.h"
 
 #if FFMA_DEBUG_ALLOCS_FREES == 1
 #include <stdio.h>
@@ -133,61 +134,6 @@ void ffma_debug_allocs_frees_end() {
 }
 #endif
 
-ffma_t **ffma_thread_cache_init() {
-    ffma_t **thread_ffmas = (ffma_t**)xalloc_alloc_zero(
-            FFMA_PREDEFINED_OBJECT_SIZES_COUNT * (sizeof(ffma_t*) + 1));
-
-    for(int i = 0; i < FFMA_PREDEFINED_OBJECT_SIZES_COUNT; i++) {
-        uint32_t object_size = ffma_predefined_object_sizes[i];
-        thread_ffmas[ffma_index_by_object_size(object_size)] = ffma_init(object_size);
-    }
-
-    return thread_ffmas;
-}
-
-void ffma_thread_cache_free(void *data) {
-    ffma_t **thread_ffmas = data;
-
-    for(int i = 0; i < FFMA_PREDEFINED_OBJECT_SIZES_COUNT; i++) {
-        uint32_t object_size = ffma_predefined_object_sizes[i];
-        uint32_t thread_ffmas_index = ffma_index_by_object_size(object_size);
-        ffma_free(thread_ffmas[thread_ffmas_index]);
-        thread_ffmas[thread_ffmas_index] = NULL;
-    }
-
-    xalloc_free(data);
-}
-
-ffma_t** ffma_thread_cache_get() {
-    ffma_t **thread_ffmas = pthread_getspecific(ffma_thread_cache_key);
-    return thread_ffmas;
-}
-
-void ffma_thread_cache_set(
-        ffma_t** ffmas) {
-    if (pthread_setspecific(ffma_thread_cache_key, ffmas) != 0) {
-        FATAL(TAG, "Unable to set the fast fixed memory allocator thread cache");
-    }
-}
-
-bool ffma_thread_cache_has() {
-    return ffma_thread_cache_get() != NULL;
-}
-
-void ffma_enable(
-        bool enable) {
-    if (enable && unlikely(!ffma_thread_cache_has())) {
-        ffma_thread_cache_set(ffma_thread_cache_init());
-    }
-
-    ffma_enabled = enable;
-    MEMORY_FENCE_STORE();
-}
-
-bool ffma_is_enabled() {
-    MEMORY_FENCE_LOAD();
-    return ffma_enabled;
-}
 
 uint8_t ffma_index_by_object_size(
         size_t object_size) {
